Add an (optional) context for commands
Commands can now register as "contextual", which makes them receive the
context as their first argument.
The context contains a timestamp (seconds since epoch), the network name
(introduced to the pladder-irc config file by this commit), the channel
the command was issued in, the nick of the user who issued the command,
and the name of the command as given.
The command in the context replaces the command argument that was
previously sent to those commands which registered as "regex".