Split mega-query into smaller queries
The whole thing with first finding the ID of the random entry and then
fetching it shouldn't improve performance anymore. Previously we only
looked at the rowid in the common table expression, but nowadays we
look at the score too. If only the rowid was inspected, then only the
index (which is already in RAM) was needed, and the table could be
left on disk.