Make authentication config more fine-grained
This allows for other authentication systems than Q to be added in the
future, as well as using a different auth user name from nick.

This commit breaks the config format, so the IRC config must be edited
if auth is used. Only raek is known to use it, so it should be fine.