Move pladdble config to pladder-bot/ and add certfile path default
Now the config file resides in ~/.config/pladder-bot/pladdble.json to
match other plugins that execute inside the pladder-bot
process. (Later we will probably extract pladdble to its own
process. Then we'll revert parts of this change.)

The certfile path now defaults to ~/.config/pladder-bot/pladdble.pem
if not set in the config file. If set in the config file, the certfile
path now needs to be absolute. It is not expected that this option
will be used much and this change was made to make the code smaller.