upgrade: Remove need for local clone of the repo
Pip can install from Git repositories over HTTPS itself, so the
upgrade script does not need to do this. The last remnats of the
upgrade script were rewritten in Python and included in the upgrade
server instead. This also solves the bug where the upgrade.sh script
was rewritten while it was executing, which sometimes lead to
problems.